pyproject.toml
[project]
name = "workspacex"
version = "0.1.0"
description = ""
authors = [
    {name = "wuhulala",email = "370031044@qq.com"}
]
readme = "README.md"
dependencies = [
    "pydantic (>=2.11.5,<3.0.0)",
    "mistune (>=3.1.3,<4.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "aiohttp (>=3.12.13,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.86.0,<2.0.0)"
]
requires-python = ">=3.10,<=3.13"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "workspacex", from = "src" },
    { include = "examples", from = "src" }

]

[tool.poetry.dependencies]
boto3 = "^1.18.0"  # 根据需要选择合适的版本